Refactor navigation imports to use centralized constants in Home, About, and Pricing pages
